requirements.txt
requests
python-dotenv
openai
streamlit
os